Add files via upload
This is an OCR that uses docTR library which is a ResNet 50 model. Has high accuracy even without a proper image preprocessing. The confidence score of each words can also be mentioned 

only drawback is the time it takes to extract an invoice or a document (In the above code it took ~1 min to do the extraction